requirements.txt
python-dotenv~=1.1.0
SQLAlchemy~=2.0.40
sqlparse~=0.5.3
langchain-google-genai~=2.1.4
langchain-core~=0.3.59
Flask~=3.1.0